+++ Tacta.EventStore/Projector/ProjectionProcessor.cs
@@ -16,7 +16,7 @@ namespace Tacta.EventStore.Projector
         private readonly IEventStoreRepository _eventStoreRepository;
         private readonly AsyncRetryPolicy _retryPolicy;
         private bool _isInitialized;
-        private int _pivot;
+        private long _pivot;
         
 
         public ProjectionProcessor(IEnumerable<IProjection> projections, IEventStoreRepository eventStoreRepository)
