+++ Tacta.EventStore/Repository/IProjectionRepository.cs
@@ -4,6 +4,6 @@ namespace Tacta.EventStore.Repository
 {
     public interface IProjectionRepository
     {
-        Task<int> GetSequenceAsync();
+        Task<long> GetSequenceAsync();
     }
 }
